add check if parameter

+++ restart-clean.sh
@@ -1,6 +1,13 @@
 #!/bin/sh
 # Stop containers, remove them and start afresh
 # give docker-compose file as param
+
+if [ -z "$1" ]
+then
+  echo "Error: missing yaml file as parameter. Exiting."
+  exit
+fi
+
 docker-compose -f $1 stop
 docker-compose -f $1 rm
 docker network rm zabbixnet
